+++ server/src/main/kotlin/org/kryptonmc/krypton/packet/in/play/PacketInTabComplete.kt
@@ -7,14 +7,11 @@ import org.kryptonmc.krypton.packet.state.PlayPacket
 
 class PacketInTabComplete : PlayPacket(0x06) {
 
-    var id: Int = -1
-        private set
-
-    lateinit var text: String
-        private set
+    var id = -1; private set
+    lateinit var command: String private set
 
     override fun read(buf: ByteBuf) {
         id = buf.readVarInt()
-        text = buf.readString(32500)
+        command = buf.readString(32500)
     }
 }